RS-454: Run gke-upgrade-test under OpenShift CI (#1103)

+++ qa-tests-backend/src/test/groovy/BaseSpecification.groovy
@@ -143,7 +143,9 @@ class BaseSpecification extends Specification {
             BaseService.setUseClientCert(false)
             withRetry(30, 1) {
                 services.ApiTokenService.revokeToken(tokenResp.metadata.id)
-                RoleService.deleteRole(testRole.name)
+                if (testRole) {
+                    RoleService.deleteRole(testRole.name)
+                }
             }
         }
